queue info discord log

+++ discordBot/discordBotManager.js
@@ -11,7 +11,8 @@ import { SanitizeDiscordLog } from '../utils/string.js';
 dotenv.config();
 
 const token = process.env.TOKEN;
-const logChannelId = process.env.DISPUTE_CHANNEL_ID
+const logChannelId = process.env.DISPUTE_CHANNEL_ID;
+const queueChannelId = process.env.QUEUE_CHANNEL_ID;
 const websiteURL = process.env.URL;
 const port = process.env.PORT;
 
@@ -19,7 +20,8 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const client = new Client({ intents: [GatewayIntentBits.Guilds] });
 
-var channel;
+var logChannel;
+var queueChannel;
 
 client.commands = new Collection();
 
@@ -27,7 +29,7 @@ var suspiciousActionsList = [];
 
 var disputeMessageId;
 var suspiciousActionsMessageId;
-
+var queueMessageId;
 
 async function GetCommands(){
 	var folderPath = path.join(__dirname, 'commands');
@@ -61,29 +63,44 @@ async function GetCommands(){
 
 client.once(Events.ClientReady, readyClient => {
 	console.log(`Discord bot logged in as ${readyClient.user.tag}`);
-	SetLogChannel();
+	SetChannels();
 });
 
-async function SetLogChannel(){
+async function SetChannels(){
 	try {
-		channel = await client.channels.fetch(logChannelId);
+		logChannel = await client.channels.fetch(logChannelId);
+		queueChannel = await client.channels.fetch(queueChannelId);
 
 		const suspiciousActionsEmbed = BuildSimpleEmbed('Suspicious actions:', 'No suspicious actions reported yet.', ' ');
 
-		const suspiciousActionsMessage = await channel.send({ embeds: [suspiciousActionsEmbed] });
+		const suspiciousActionsMessage = await logChannel.send({ embeds: [suspiciousActionsEmbed] });
 		suspiciousActionsMessageId = suspiciousActionsMessage.id;
 
 		const disputeEmbed = {
-		color: embedColor,
-		title: 'Current disputes:',
-		fields: [{name: 'There are currently no disputes.', value: ' '}],
-		timestamp: new Date().toISOString(),
+			color: embedColor,
+			title: 'Current disputes:',
+			fields: [{name: 'There are currently no disputes.', value: ' '}],
+			timestamp: new Date().toISOString(),
 		};
 
-		const disputeMessage = await channel.send({ embeds: [disputeEmbed] });
+		const disputeMessage = await logChannel.send({ embeds: [disputeEmbed] });
 		disputeMessageId = disputeMessage.id;
+
+		const queueEmbed = GetQueueEmbed([0, 0]);
+
+		const queueMessage = await queueChannel.send({ embeds: [queueEmbed] });
+		queueMessageId = queueMessage.id;
+
 	} catch(error){
 		console.log(error);
+	}
+}
+
+function GetQueueEmbed(queueInfo){
+	return {
+		color: embedColor,
+		title: 'Players in queue:',
+		fields: [{name: 'Casual', value: queueInfo[0]}, {name: 'Ranked', value: queueInfo[1]}],
 	}
 }
 
@@ -129,7 +146,7 @@ export async function StartDiscordBot(){
 
 export async function SendDisputeMessage(matchDisputes, sendNewMessage){
 	try {
-		if (!channel) return;
+		if (!logChannel) return;
 
 		//build embed
 		let fields = [];
@@ -145,17 +162,35 @@ export async function SendDisputeMessage(matchDisputes, sendNewMessage){
 		if (matchDisputes.length == 0) fields.push({name: 'There are currently no disputes.', value: ' '});
 
 		const disputeEmbed = {
-		color: embedColor,
-		title: 'Current disputes:',
-		fields: fields,
-		timestamp: new Date().toISOString(),
+			color: embedColor,
+			title: 'Current disputes:',
+			fields: fields,
+			timestamp: new Date().toISOString(),
 		};
 
-		const previousMessage = await channel.messages.fetch(disputeMessageId);
+		const previousMessage = await logChannel.messages.fetch(disputeMessageId);
 		if (previousMessage) await previousMessage.edit({ embeds: [disputeEmbed] });
 
 		if (sendNewMessage){
-			const tempMessage = await channel.send('ping');
+			const tempMessage = await logChannel.send('ping');
+			await tempMessage.delete();
+		}
+	} catch (error){
+		console.log(error);
+	}
+}
+
+export async function SendQueueInfo(queueInfo, sendNewMessage){
+	try {
+		if (!queueChannel) return;
+
+		const queueEmbed = GetQueueEmbed(queueInfo);
+
+		const previousMessage = await queueChannel.messages.fetch(queueMessageId);
+		if (previousMessage) await previousMessage.edit({ embeds: [queueEmbed] });
+
+		if (sendNewMessage){
+			const tempMessage = await queueChannel.send('ping');
 			await tempMessage.delete();
 		}
 	} catch (error){
@@ -175,7 +210,7 @@ export async function SendNewSuspiciousAction(suspiciousAction){
 	if (suspiciousActionsList.length > 25) suspiciousActionsList.shift();
 
 	try {
-		if (!channel) return;
+		if (!logChannel) return;
 
 		//build embed
 		let fields = [];
@@ -197,7 +232,7 @@ export async function SendNewSuspiciousAction(suspiciousAction){
 		fields: fields,
 		};
 
-		const previousMessage = await channel.messages.fetch(suspiciousActionsMessageId);
+		const previousMessage = await logChannel.messages.fetch(suspiciousActionsMessageId);
 		if (previousMessage) await previousMessage.edit({ embeds: [suspiciousActionsEmbed] });
 	} catch (error){
 		console.log(error);
